requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
filelock==3.19.1
fsspec==2025.9.0
idna==3.11
Jinja2==3.1.6
lingua-language-detector==2.1.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
pandas==2.3.3
pillow==11.3.0
pybind11==3.0.1
PyMuPDF==1.26.6
pymupdf4llm==0.2.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
torch==2.9.1+cu128
torchvision==0.24.1+cu128
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
